requirements.txt
python-dotenv==1.0.1

# Base perf is faster than `httpx`
aiohttp==3.9.5

# Standard Connector that provided from MySQL
mysql-connector-python==8.4.0
